rm VMAdapter
